feat: refactor command and query namespaces for better organization and clarity #67
