+++ Services/Order/src/Order.Application/Queries/GetOrderById/GetOrderByIdQuery.cs
@@ -1,6 +1,7 @@
 using FluentResults;
 using MediatR;
 using Order.Application.Responses;
-namespace Order.Application.Queries;
+
+namespace Order.Application.Queries.GetOrderById;
 
 public record GetOrderByIdQuery(int Id) : IRequest<Result<GetOrderComposeResponse>>;
