+++ Services/Order/tests/Order.UnitTests/Application/Commands/CreateOrderTests.cs
@@ -1,12 +1,11 @@
+using BuildingBlocks.Contracts.Events;
+using BuildingBlocks.Contracts.MongoEvents;
 using BuildingBlocks.Security.Context;
 using FluentAssertions;
 using Moq;
-using Order.Application.Commands.Handlers;
-using BuildingBlocks.Contracts.Events;
-using BuildingBlocks.Contracts.MongoEvents;
+using Order.Application.Commands.CreateOrder;
 using Order.Domain.Interfaces;
 using Order.Domain.Models;
-using Order.Application.Commands.CreateOrder;
 namespace Order.UnitTests.Application.Commands;
 
 public class CreateOrderTests
