+++ Services/Order/tests/Order.UnitTests/Application/Commands/PaymentUpdatedCommandTests.cs
@@ -3,8 +3,7 @@ using BuildingBlocks.Messaging;
 using FluentAssertions;
 using MediatR;
 using Moq;
-using Order.Application.Commands;
-using Order.Application.Commands.Handlers;
+using Order.Application.Commands.PaymentUpdated;
 using Order.Application.Interfaces;
 using Order.Domain.Interfaces;
 using Order.Domain.Models;
