+++ Services/Stock/src/Stock.Api/Controllers/ProductController.cs
@@ -4,7 +4,8 @@ using Microsoft.AspNetCore.Mvc;
 using Stock.Application.Commands.CreateProduct;
 using Stock.Application.Commands.UpdateProduct;
 using Stock.Application.Commands.UpdateStock;
-using Stock.Application.Queries;
+using Stock.Application.Queries.GetAllProducts;
+using Stock.Application.Queries.GetProductById;
 using System.ComponentModel.DataAnnotations;
 namespace Stock.Api.Controllers;
