+++ Services/Stock/src/Stock.Application/Queries/GetAllProducts/GetAllProductsQuery.cs
@@ -1,6 +1,6 @@
 using FluentResults;
 using MediatR;
 using Stock.Application.Responses;
-namespace Stock.Application.Queries;
+namespace Stock.Application.Queries.GetAllProducts;
 
 public record GetAllProductsQuery() : IRequest<Result<IEnumerable<GetProductResponse>>>;
