+++ Services/Stock/src/Stock.Application/Queries/GetProductById/GetProductByIdQuery.cs
@@ -1,6 +1,6 @@
 using FluentResults;
 using MediatR;
 using Stock.Application.Responses;
-namespace Stock.Application.Queries;
+namespace Stock.Application.Queries.GetProductById;
 
 public record GetProductByIdQuery(int Id) : IRequest<Result<GetProductResponse>>;
