+++ Services/Stock/tests/Stock.UnitTests/Application/Commands/OrderFailedTests.cs
@@ -2,8 +2,7 @@ using BuildingBlocks.Contracts;
 using FluentAssertions;
 using MediatR;
 using Moq;
-using Stock.Application.Commands;
-using Stock.Application.Commands.Handlers;
+using Stock.Application.Commands.OrderFailed;
 using Stock.Application.Interfaces;
 using Stock.Domain.Interfaces;
 using Stock.Domain.Models;
