+++ Services/Stock/tests/Stock.UnitTests/Application/Commands/OrderRequestTests.cs
@@ -4,8 +4,7 @@ using BuildingBlocks.Messaging;
 using FluentAssertions;
 using MediatR;
 using Moq;
-using Stock.Application.Commands;
-using Stock.Application.Commands.Handlers;
+using Stock.Application.Commands.OrderRequest;
 using Stock.Application.Interfaces;
 using Stock.Domain.Interfaces;
 using Stock.Domain.Models;
