+++ Services/User/src/User.Application/Queries/GetAllUsers/GetAllUsersQuery.cs
@@ -1,6 +1,6 @@
 using FluentResults;
 using MediatR;
 using User.Application.Responses;
-namespace User.Application.Queries;
+namespace User.Application.Queries.GetAllUsers;
 
 public record GetAllUsersQuery() : IRequest<Result<IEnumerable<GetUserResponse>>>;
