+++ Services/User/src/User.Application/Queries/GetUsersById/GetUserByIdQuery.cs
@@ -1,6 +1,7 @@
 using FluentResults;
 using MediatR;
 using User.Application.Responses;
-namespace User.Application.Queries;
+
+namespace User.Application.Queries.GetUsersById;
 
 public record GetUserByIdQuery(int Id) : IRequest<Result<GetUserResponse>>;
