+++ Services/User/src/User.Application/Queries/GetUsersById/GetUserByIdQueryHandler.cs
@@ -3,28 +3,33 @@ using MediatR;
 using Serilog;
 using User.Application.Interfaces;
 using User.Application.Responses;
-namespace User.Application.Queries.Handlers;
+
+namespace User.Application.Queries.GetUsersById;
 
 public class GetUserByIdQueryHandler : IRequestHandler<GetUserByIdQuery, Result<GetUserResponse>>
 {
     private readonly IUserReadService _userService;
     private readonly ILogger _logger;
+
     public GetUserByIdQueryHandler(IUserReadService userService)
     {
         _userService = userService;
         _logger = Log.ForContext<GetUserByIdQueryHandler>();
     }
+
     public async Task<Result<GetUserResponse>> Handle(GetUserByIdQuery request, CancellationToken cancellationToken)
     {
         try
         {
             _logger.Information("[INFO] Handling {EventName} for UserId: {UserId}", nameof(GetUserByIdQuery), request.Id);
+
             var user = await _userService.GetByIdAsync(request.Id);
             if (user is null)
             {
                 _logger.Warning("[WARN] User with ID {UserId} not found in the repository", request.Id);
                 return Result.Fail<GetUserResponse>($"User with ID {request.Id} not found.");
             }
+
             var response = new GetUserResponse(
                 user.Id,
                 user.Username,
@@ -33,7 +38,9 @@ public class GetUserByIdQueryHandler : IRequestHandler<GetUserByIdQuery, Result<
                 user.CreatedAt,
                 user.UpdatedAt
             );
+
             _logger.Information("[INFO] Successfully fetched and mapped user with ID {UserId}", request.Id);
+
             return Result.Ok(response);
         }
         catch (Exception ex)
